refactor(app): split the hosting-platform emitters into per-provider modules
env-format.ts had grown to hold five providers' quirks alongside the generic
formats. Each provider now owns a module under shared/deploy/ — render, fly,
railway, koyeb, coolify — so Koyeb's volume caveats and Fly's secrets handling
sit next to nothing else.

The shared vocabulary (EnvEntry/EmitOptions, the per-syntax quoting rules, the
data-mount and health-path constants) moves down into env-format-base.ts. That
is what breaks the cycle: the provider modules depend on the base, env-format.ts
depends on both, and nothing depends back up. env-format.ts stays the single
public entry point and re-exports everything, so #shared/env-format callers are
untouched and the generated manifests come out byte-identical.

Kept as pure (entries, opts) => string functions selected through the existing
ENV_OUTPUT_FORMATS registry rather than factory classes: the emitters hold no
state, so a class per provider would add a constructor and an instance for
nothing, and would break symmetry with the eight generic emitters. The registry
already provides the dispatch a factory would.

Cross-module imports inside shared/ use #shared/... specifiers, which resolve
identically in Vite, Vitest and plain Node (via the package.json imports map) —
relative paths would have forced .ts extensions that the Bundler moduleResolution
rejects.
