+++ application/shared/deploy/coolify.ts
@@ -1,0 +1,54 @@
+import {
+  commentBlock,
+  DATA_MOUNT,
+  DEFAULT_IMAGE,
+  DEFAULT_NAME,
+  DEFAULT_PORT,
+  HEALTH_PATH,
+  magicIdentifier,
+  type EmitOptions,
+  type EnvEntry,
+} from '#shared/env-format-base';
+
+/**
+ * A Compose stack for the self-hostable PaaS layer — Coolify and Dokploy both
+ * consume one. Uses Coolify's magic variables so the domain and the secrets are
+ * generated by the platform rather than pasted in by hand.
+ */
+export function emitCoolifyCompose(entries: readonly EnvEntry[], opts: EmitOptions = {}): string {
+  const name = opts.name ?? DEFAULT_NAME;
+  const port = opts.port ?? DEFAULT_PORT;
+  const fqdnVar = `SERVICE_FQDN_${magicIdentifier(name)}_${port}`;
+  const env: string[] = [`      - ${fqdnVar}`];
+  for (const entry of entries) {
+    if (entry.comment) env.push(`      # ${entry.comment}`);
+    if (entry.platformValue === 'generate') {
+      env.push(`      - ${entry.name}=\${SERVICE_PASSWORD_64_${magicIdentifier(entry.name)}}`);
+    } else if (entry.platformValue === 'url') {
+      env.push(`      - ${entry.name}=\${${fqdnVar}}`);
+    } else {
+      env.push(`      - ${entry.name}=${entry.value}`);
+    }
+  }
+  return (
+    commentBlock(opts.header) +
+    `services:
+  ${name}:
+    image: ${opts.image ?? DEFAULT_IMAGE}
+    environment:
+${env.join('\n')}
+    volumes:
+      - ${name}-data:${DATA_MOUNT}
+    healthcheck:
+      test: ['CMD', 'wget', '-qO', '/dev/null', 'http://127.0.0.1:${port}${HEALTH_PATH}']
+      interval: 30s
+      timeout: 5s
+      retries: 3
+      start_period: 20s
+    restart: unless-stopped
+
+volumes:
+  ${name}-data:
+`
+  );
+}
